// common/types.ts
export type UserProps = {
	id?: string | undefined;
	email?: string | undefined;
	password?: string | undefined;
	rememberMe?: boolean | undefined;
};

export type CollaboratorProps = {
	id?: string | undefined;
	name?: string | undefined;
	cpf?: string | undefined;
	tel?: string | undefined;
	email?: string | undefined;
};